requirements.txt
black==24.3.0
isort==5.13.2
flake8==7.1.0
mypy==1.10.0
pytest==8.2.0
bandit==1.7.9
safety==3.2.0
pyyaml==6.0.2
azure-functions
azure-identity
azure-storage-blob
opentelemetry-sdk
requests
python-dotenv
pinecone>=5.0.0
openai
pyyaml
fastapi
uvicorn

